Synch questions working fine. Continue button removed. Fixed #64

+++ client/DotaAsker/ChangeAvatarViewController.m
@@ -42,7 +42,7 @@
     
     // [4]
     NSUInteger row = [_avatarNamesArray indexOfObject:[[Player instance] avatarImageName]];
-    NSLog(@"Found avatar in array. Index: %lu", row);
+    NSLog(@"Found avatar in array. Index: %lu", (unsigned long)row);
     NSIndexPath* path = [NSIndexPath indexPathForRow:row inSection:0];
     [self.collectionView selectItemAtIndexPath:path animated:YES scrollPosition:UICollectionViewScrollPositionTop];
     [self collectionView:self.collectionView didSelectItemAtIndexPath:path];
