+++ client/DotaAsker/MatchViewController.h
@@ -16,7 +16,7 @@
 <UITableViewDataSource, UITableViewDelegate, RoundViewDelegate>
 
 @property (assign, nonatomic) NSUInteger buttonState;
-@property (assign, nonatomic) NSUInteger matchID;
+@property (assign, nonatomic) long long matchID;
 @property (strong, nonatomic) MatchViewModel *matchViewModel;
 @property (strong, nonatomic) IBOutlet UITableView* tableView;
